requirements_v1.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
antlr4-python3-runtime==4.9.3
anyio==4.3.0
appdirs==1.4.4
asteroid-filterbanks==0.4.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
braceexpand==0.1.7
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.8.2
comm==0.2.2
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.9
decorator==5.1.1
Distance==0.1.3
docker-pycreds==0.4.0
docopt==0.6.2
editdistance==0.8.1
einops==0.7.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.0
filelock==3.13.1
flatbuffers==24.3.7
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.3.1
g2p-en==2.1.0
gevent==24.2.1
geventhttpclient==2.0.2
gitdb==4.0.11
GitPython==3.1.42
greenlet==3.0.3
grpcio==1.62.1
h11==0.14.0
huggingface-hub==0.21.4
humanfriendly==10.0
hydra-core==1.3.2
HyperPyYAML==1.2.2
idna==3.6
inflect==7.0.0
ipython==8.22.2
ipywidgets==8.1.2
jedi==0.19.1
Jinja2==3.1.3
jiwer==3.0.3
joblib==1.3.2
julius==0.2.7
jupyterlab_widgets==3.0.10
kaldi-python-io==1.2.2
kaldiio==2.18.0
kiwisolver==1.4.5
lazy_loader==0.3
Levenshtein==0.25.0
librosa==0.10.1
lightning==2.2.1
lightning-utilities==0.11.0
llvmlite==0.42.0
loguru==0.7.2
Mako==1.3.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
nemo-toolkit==1.21.0
networkx==3.2.1
nltk==3.8.1
numba==0.59.1
numpy==1.23.5
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.17.1
optuna==3.6.0
packaging==24.0
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
plac==1.4.3
platformdirs==4.2.0
pooch==1.8.1
primePy==1.3
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyannote.audio==3.1.1
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pybind11==2.11.1
pycparser==2.21
pydantic==1.10.14
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-multipart==0.0.9
python-rapidjson==1.16
pytorch-lightning==2.0.7
pytorch-metric-learning==2.4.1
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.6.2
regex==2023.12.25
requests==2.31.0
rich==13.7.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
sacremoses==0.1.1
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
semver==3.0.2
sentencepiece==0.2.0
sentry-sdk==1.43.0
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
sox==1.5.0
soxr==0.3.7
speechbrain==1.0.0
SQLAlchemy==2.0.28
stack-data==0.6.3
starlette==0.36.3
sympy==1.12
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.4.0
text-unidecode==1.3
texterrors==0.4.4
threadpoolctl==3.4.0
tokenizers==0.15.2
torch==2.2.1
torch-audiomentations==0.11.1
torch-pitch-shift==1.2.4
torchaudio==2.2.1
torchmetrics==1.3.2
tqdm==4.66.2
traitlets==5.14.2
transformers==4.37.1
tritonclient==2.36.0
typer==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
wandb==0.16.4
wcwidth==0.2.13
webdataset==0.1.62
Werkzeug==3.0.1
wget==3.2
widgetsnbextension==4.0.10
wrapt==1.16.0
yarl==1.9.4
youtokentome==1.0.6
zope.event==5.0
zope.interface==6.2
